src/main/java/org/kpmp/eridanus/notifications/StateChangeEvent.java
package org.kpmp.eridanus.notifications;

class StateChangeEvent {

	private String state;
	private String packageId;
	private String origin;
	private String codicil;

	public String getState() {
		return state;
	}

	public void setState(String state) {
		this.state = state;
	}

	public String getPackageId() {
		return packageId;
	}

	public void setPackageId(String packageId) {
		this.packageId = packageId;
	}

	public String getOrigin() {
		return origin;
	}

	public void setOrigin(String origin) {
		this.origin = origin;
	}

	public String getCodicil() {
		return codicil;
	}

	public void setCodicil(String codicil) {
		this.codicil = codicil;
	}

}
